% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcall_classes.R
\name{Clustering}
\alias{Clustering}
\title{Clustering R6 Class (Internal)}
\description{
Internal R6 class for automatic legislator clustering. Users should use bcall_auto() instead.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{N}}{Distance metric (1=Manhattan, 2=Euclidean)}

\item{\code{X}}{Rollcall data matrix}

\item{\code{clusters}}{List of cluster assignments}

\item{\code{clustering}}{Final clustering dataframe}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Clustering-new}{\code{Clustering$new()}}
\item \href{#method-Clustering-initialize_clusters}{\code{Clustering$initialize_clusters()}}
\item \href{#method-Clustering-iterate_clustering}{\code{Clustering$iterate_clustering()}}
\item \href{#method-Clustering-iteration}{\code{Clustering$iteration()}}
\item \href{#method-Clustering-get_distances}{\code{Clustering$get_distances()}}
\item \href{#method-Clustering-assign_clusters}{\code{Clustering$assign_clusters()}}
\item \href{#method-Clustering-create_clustering_dataframe}{\code{Clustering$create_clustering_dataframe()}}
\item \href{#method-Clustering-reclassify_legislators}{\code{Clustering$reclassify_legislators()}}
\item \href{#method-Clustering-clone}{\code{Clustering$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-new"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-new}{}}}
\subsection{Method \code{new()}}{
Initialize Clustering object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$new(rollcalls, N = 1, pivot = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rollcalls}}{Data frame with rollcall data}

\item{\code{N}}{Integer, distance metric (1 or 2)}

\item{\code{pivot}}{Character, pivot legislator name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-initialize_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-initialize_clusters}{}}}
\subsection{Method \code{initialize_clusters()}}{
Initialize clusters with farthest legislators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$initialize_clusters(pivot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pivot}}{Character, pivot legislator name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-iterate_clustering"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-iterate_clustering}{}}}
\subsection{Method \code{iterate_clustering()}}{
Iterate clustering assignment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$iterate_clustering()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-iteration"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-iteration}{}}}
\subsection{Method \code{iteration()}}{
Single clustering iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$iteration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-get_distances"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-get_distances}{}}}
\subsection{Method \code{get_distances()}}{
Calculate distances between tensors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$get_distances(tensor1, tensor2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tensor1}}{First data matrix}

\item{\code{tensor2}}{Second data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-assign_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-assign_clusters}{}}}
\subsection{Method \code{assign_clusters()}}{
Assign left and right clusters based on pivot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$assign_clusters(pivot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pivot}}{Character, pivot legislator name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-create_clustering_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-create_clustering_dataframe}{}}}
\subsection{Method \code{create_clustering_dataframe()}}{
Create clustering dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$create_clustering_dataframe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-reclassify_legislators"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-reclassify_legislators}{}}}
\subsection{Method \code{reclassify_legislators()}}{
Reclassify legislators to optimal clusters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$reclassify_legislators()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Clustering-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Clustering-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Clustering$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
